# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "REST API Documentation for yaca-api",
        "version": "1",
        "description": "REST API made for real state project by YACA"
    },
    "servers": [
        {
            "url": "https://pruebas.dani9oo.dev/",
            "description": "Development API"
        }
    ],
    "tags": [
        {
            "name": "User",
            "description": "User data operations"
        }
    ],
    "paths": {
        "/api/rest/v1/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Route used to register new user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "birthdate": {
                                        "type": "string"
                                    },
                                    "telephone": {
                                        "type": "string"
                                    },
                                    "dialcode": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Whether or not the operation was successful"
                                        },
                                        "result": {
                                            "type": "object",
                                            "properties": {
                                                "birthdate": {
                                                    "type": "string"
                                                },
                                                "lastname": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "telephone": {
                                                    "type": "string"
                                                },
                                                "dialcode": {
                                                    "type": "string"
                                                },
                                                "username": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "password": {
                                                    "type": "string"
                                                },
                                                "role": {
                                                    "type": "integer"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                },
                                                "_id": {
                                                    "type": "string"
                                                },
                                                "createdAt": {
                                                    "type": "string"
                                                }
                                            },
                                            "description": "Result message, including MongoId of object created"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rest/v1/user/{_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/MongoId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Whether or not the operation was successful"
                                        },
                                        "result": {
                                            "type": "object",
                                            "properties": {
                                                "birthdate": {
                                                    "type": "string"
                                                },
                                                "lastname": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "telephone": {
                                                    "type": "string"
                                                },
                                                "dialcode": {
                                                    "type": "string"
                                                },
                                                "username": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "password": {
                                                    "type": "string"
                                                },
                                                "role": {
                                                    "type": "integer"
                                                },
                                                "status": {
                                                    "type": "integer"
                                                },
                                                "_id": {
                                                    "type": "string"
                                                },
                                                "createdAt": {
                                                    "type": "string"
                                                }
                                            },
                                            "description": "Result message, including MongoId of object created"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rest/v1/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Route used to login user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email":{
                                        "type":"string"
                                    },
                                    "password":{
                                        "type":"string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "succes": {
                                            "type": "boolean"
                                                 }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "MongoId": {
                "name": "_id",
                "in": "path",
                "schema": {
                    "type": "string",
                    "format": "MongoId"
                },
                "required": true
            }
        }
    }
}